Try new file structure, add neomutt/bash_profile
Try out a new file structure more akin to the actual home dir from
which these files are (mostly) sourced. Separate .mutt and the neomutt
dir to help determine whether this and the current list of files is
ideal, and add bash_profile to fully round out the interactive shell
environment.