Provide my Myth the most bleeding-edge setup
Significantly extend and alter .bashrc and .bash_profile, with the
side effect of partial and partially externally sourced contributions
lying scattered around ("graveyard" status). Add playvid and profrc
(which also contains aliases), along with an unbroken but highly
stripped-down firefox user.js, which is merely the other end of the
compatibility spectrum and requires integration with arkenfox.